Move toBeReset and willBeConnected logic to gameboardtiles
