+++ client/src/features/game/GameBoard.tsx
@@ -54,7 +54,6 @@ const GameBoard: React.FC<GameBoardProps> = ({ id, game, userIndex }) => {
   const selectedWord = useAppSelector((state) =>
     getSelectedWordByGameId(state, id)
   );
-  const selection = useAppSelector(getOrderedTileSelectionCoords);
   const replayLetters = useAppSelector(
     (state) => state.game.replay.move?.letters
   );
@@ -99,28 +98,11 @@ const GameBoard: React.FC<GameBoardProps> = ({ id, game, userIndex }) => {
   const portal = useRef<HTMLDivElement>(null);
 
   const currentMove = useAppSelector(getTileSelectionInParsedHexCoords);
-  const replayMove = useAppSelector((state) => state.game.replay.move);
+  // const replayMove = useAppSelector((state) => state.game.replay.move);
   const replayTiles = useAppSelector(
     getHightlightedCoordsForCurrentReplayState
   );
 
-  const tilesThatWillBeCaptured = useMemo(() => {
-    const willBeCaptured = {};
-    currentMove.forEach((coord) => {
-      willBeCaptured[`${coord.q},${coord.r}`] = willConnectToTerritory(
-        game.grid,
-        replayMove?.coords ?? currentMove,
-        coord,
-        game.turn
-      );
-    });
-    return willBeCaptured;
-  }, [currentMove, replayMove?.coords, game.grid, game.turn]);
-
-  const tilesThatWillBeReset = useMemo(() => {
-    const cells = getCellsToBeReset(game.grid, replayMove?.coords ?? currentMove, game.turn);
-    return R.groupBy((cell: Cell) => `${cell.q},${cell.r}`, cells);
-  }, [game.grid, game.turn, currentMove, replayMove?.coords]);
 
   return (
     <>
@@ -241,11 +223,9 @@ const GameBoard: React.FC<GameBoardProps> = ({ id, game, userIndex }) => {
               !submitting && !game.gameOver && userIndex === game.turn
             }
             position={[0, -7, 0]}
-            selection={replayTiles ?? selection}
+            selection={replayTiles ?? currentMove}
             currentTurn={game.turn}
             onToggleTile={(coord) => dispatch(toggleTileSelected(coord))}
-            tilesThatWillBeCaptured={tilesThatWillBeCaptured}
-            tilesThatWillBeReset={tilesThatWillBeReset}
           />
         </React.Suspense>
       </Canvas>
